requirements.txt
python==3.11
pytorch==2.1.0
cuda==12.1